add API on payment module fix the margin bug of self info

+++ app/src/main/java/com/iplay/feastbooking/gson/orderDetail/OrderDetail.java
@@ -91,7 +91,7 @@ public class OrderDetail implements Serializable {
     }
 
     public static class OrderPayment implements Serializable{
-        public int amountPaid;
+        public double amountPaid;
 
         public String approvalStatus;
